Add channels privileges and fix 500 errors

+++ src/main/java/udes/chat_api/gateway/ChannelGateway.java
@@ -3,6 +3,7 @@ package udes.chat_api.gateway;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import udes.chat_api.channels.Channel;
+import udes.chat_api.channels.ChannelRepository;
 import udes.chat_api.channels.ChannelService;
 
 import java.util.List;
@@ -12,6 +13,10 @@ public class ChannelGateway
 {
     @Autowired
     private ChannelService channelService;
+    @Autowired
+    private MainGateway mainGateway;
+    @Autowired
+    private ChannelRepository channelRepository;
 
     public List<Channel> getChannelsByRoomId(int channelId)
     {
@@ -20,29 +25,33 @@ public class ChannelGateway
 
     public Channel createChannel(Channel channel)
     {
-        // Check room_privileges
-        // Return privilege error if the user does not have the required room_privileges
+        if(!mainGateway.isAdminOrModerator(channel.getRoom()))
+        {
+            return null;
+        }
 
         return channelService.createChannel(channel);
     }
 
-    public Channel getChannel(int channelId)
-    {
-        return channelService.getChannel(channelId);
-    }
-
-    public List<Channel> searchChannel(String query)
-    {
-        return channelService.searchChannel(query);
-    }
-
     public Channel updateChannel(Channel channel)
     {
+        if(!mainGateway.isAdminOrModerator(channel.getRoom()))
+        {
+            return null;
+        }
+
         return channelService.updateChannel(channel);
     }
 
     public Channel deleteChannel(int channelId)
     {
+        Channel channel = channelRepository.findByChannelIdAndIsDeletedFalse(channelId);
+
+        if(!mainGateway.isAdminOrModerator(channel.getRoom()))
+        {
+            return null;
+        }
+
         return channelService.deleteChannel(channelId);
     }
 }
